chore: fix gosec G115 lint

+++ internal/pkg/util/user/identity_unix.go
@@ -70,7 +70,7 @@ func CurrentOriginal() (*User, error) {
 	if err != nil {
 		return nil, err
 	}
-	return GetPwUID(uint32(uid))
+	return GetPwUID(uid)
 }
 
 // SetCurrentOriginal sets the current original user, used to set the
